Audio language setting, menu re-arrangement, small future-proofing
There is now an option in the audio options menu which allows the
player to toggle between English and Japanese audio.
The menus have been re-arranged, mostly as described in the
Google Doc tasks, except that I picked a different order for the
options under Single Player because I thought they made more
sense that way.
Also, OpenSakAlt is meant to take an OPTIONAL script file but this
was actually required, despite the code being able to work without
it. I just added a default value of "false" for this argument,
which makes it optional and seems to work (although I don't have
a need for OpenSakAlt here, so it's not currently used).